Added NTFS, ext3 and ext4 sample images.
All images contain several folders with nested subfolders and text files
on each folder. Each text file contains "this file is: {FILENAME}"

Each image is only 2MB to keep the repo size as short as possible as
this files are mainly to test the `m` commands.

There is a ext4_advanced.img image with all newer and advanced features
from ext4 that at this moment are incompatible with the grub driver
radare uses to access ext file systems.

The flags I had to disable were:

1. metadata_csum - Metadata checksumming for data integrity verification
2. metadata_csum_seed - Seed for metadata checksum calculation
3. 64bit - Allows filesystems larger than 2^32 blocks
4. orphan_file - Dedicated file for orphan inode tracking (newer feature)

These are all relatively modern ext4 features that the grub ext2 driver
doesn't support. The grub driver was designed primarily for ext2/3
compatibility with basic ext4 support.